Fix the return size of xb_silo_node_get_size()
In theory this could overflow.

+++ src/xb-silo-private.h
@@ -67,7 +67,7 @@ guint32		 xb_silo_get_strtab_idx		(XbSilo		*self,
 						 const gchar	*element);
 guint32		 xb_silo_get_offset_for_node	(XbSilo		*self,
 						 XbSiloNode	*n);
-guint8		 xb_silo_node_get_size		(XbSiloNode	*n);
+guint32		 xb_silo_node_get_size		(XbSiloNode	*n);
 XbSiloNode	*xb_silo_get_sroot		(XbSilo		*self);
 XbSiloNode	*xb_silo_node_get_parent	(XbSilo		*self,
 						 XbSiloNode	*n);
